+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 0.15.0
 - Events: Embed SelectRef objects instead of IDs
 - Entity with state: Edit form should not be able to created new data
+- Entity with state: Table with "Action" col only if there are non apiManaged actions 
 - fix double Docu nav
+
 
 # 0.14.0
 - Add new tenants via GUI